Add ShowCocktail component Clean RandomCocktail page
